\name{InfoCreate}
\alias{InfoCreate}
\title{Create info list function for QT}
\description{Creates info list for QT data analyses.}
\usage{
	 InfoCreate(info)
}
\arguments{\item{info}{See \link{QT}}}

\value{
  Creates info list with all required arguments.
}
\note{
  For examples see \link{QT}.
}

\author{Christoffer W. Tornoe \email{pharmacometrics@fda.hhs.gov}}
\note{DISCLAIMER: This R package does not reflect the official policy of the FDA. No official endorsement by the FDA is intended or should be inferred.}
\keyword{models}
