\name{MeanData}
\alias{MeanData}
\title{Mean data function}
\description{Creates standard analysis dataset for QTc-Time and concentration-QTc analyses.}
\usage{
	MeanData(data,info)
}

\arguments{
	\item{data}{See \link{QT}}
	\item{info}{See \link{QT}}
}

\author{Christoffer W. Tornoe <\email{pharmacometrics@fda.hhs.gov}>}
\note{DISCLAIMER: This R package does not reflect the official policy of the FDA. No official endorsement by the FDA is intended or should be inferred.}
\keyword{models}
