\name{outputGraph}
\alias{outputGraph}
\title{Function to print graphs using different devices}
\description{Creates info list for QT data analyses.}
\usage{
	 outputGraph(name,device,height,width)
}
\arguments{
	\item{name}{File name of graph.}
	\item{device}{Graphical device to print to (options: "wmf", "bmp", "jpg", "pdf", "png", or "tiff") (default is "tiff").}
	\item{height}{Height of graph.}
	\item{width}{Width of graph.}
}

\note{
  For examples see \link{QT}.
}

\author{Christoffer W. Tornoe \email{pharmacometrics@fda.hhs.gov}}
\note{DISCLAIMER: This R package does not reflect the official policy of the FDA. No official endorsement by the FDA is intended or should be inferred.}
\keyword{models}
