\name{QTconc}
\alias{QTconc}
\title{Concentration-QTc Function}
\description{Performs concentration-QTc Analyses.}
\usage{
	QTconc(info)
}

\arguments{
	\item{info}{See \link{QT}}
}
\author{Christoffer W. Tornoe <\email{pharmacometrics@fda.hhs.gov}>}
\note{DISCLAIMER: This R package does not reflect the official policy of the FDA. No official endorsement by the FDA is intended or should be inferred.}
\keyword{models}
