\name{QTcorrections}
\alias{QTcorrections}
\title{QT-RR Function}
\description{Performs QT-RR analyses.}
\usage{
	QTcorrections(data,info)
}

\arguments{
	\item{data}{See \link{QT}}
	\item{info}{See \link{QT}}
}

\author{Christoffer W. Tornoe <\email{pharmacometrics@fda.hhs.gov}>}
\note{DISCLAIMER: This R package does not reflect the official policy of the FDA. No official endorsement by the FDA is intended or should be inferred.}
\keyword{models}
