\name{QTtime}
\alias{QTtime}
\title{QT-Time Function}
\description{Performs QTc-Time Analyses.}
\usage{
	QTtime(info)
}

\arguments{
	\item{info}{See \link{QT}}
}

\author{Christoffer W. Tornoe <\email{pharmacometrics@fda.hhs.gov}>}
\note{DISCLAIMER: This R package does not reflect the official policy of the FDA. No official endorsement by the FDA is intended or should be inferred.}
\keyword{models}
