\name{qtpk}
\docType{data}
\alias{qtpk}
\title{QT dataset}
\description{Simulated QT dataset}
\keyword{datasets}
\author{Christoffer W. Tornoe <\email{pharmacometrics@fda.hhs.gov}>}
\note{DISCLAIMER: This R package does not reflect the official policy of the FDA. No official endorsement by the FDA is intended or should be inferred.}
